added a new file
One of the organization wants to deploy a new multi-tier application. The application will take live inputs from the employees and it will be hosted on a web server running on the AWS cloud. 

The development team has asked you me set up the web server and configure it to scale automatically in cases of a traffic surge, to make the application highly available. They have also asked me to take the inputs from the employees and store them securely in the database.